move the new TOML tests into their own translation unit
Adding them to toml_test.cpp pushed cc1plus into an OOM on every GCC job:
"fatal error: Killed signal terminated program cc1plus" building
toml_test.cpp.o, on gcc15, gcc16, gcc16-reflection, asan, armv7 and s390x.
Clang was unaffected.

Measured with clang, toml_test.cpp peaks at 965MB on main and the additions
took it to 1241MB. The ambiguous-nest case was 122MB of that on its own: it
instantiated a reader per level of a 24-deep variant nest. The depth cannot be
lowered to compensate, because the speculation budget saturates around depth
18 and below that the nest costs the same bounded or not, so the test stops
discriminating.

Two changes together:

- toml_test.cpp goes back to exactly what it was, and the new suites move to
  toml_reader_test. Later TOML reader tests now have somewhere to go that is
  not the file already at the limit.

- The nest recurses through a std::vector member rather than through a
  template parameter, so its depth costs runtime nesting instead of one
  instantiated reader per level: three types instead of fifty, and 34MB of
  compile memory instead of 185MB. It discriminates harder for it -- 41s
  without the budget fix against 0.56s with it, where the templated form at
  the same depth was 10s.

toml_reader_test peaks at 366MB, against 990MB for toml_test, 1.76GB for
yaml_test and 2.57GB for json_test. That matters beyond the one file: these
jobs build several heavy translation units at once, and the asan runner had
gone over on an unrelated networking test rather than on this one.
