+++ tests/CMakeLists.txt
@@ -105,6 +105,7 @@ add_subdirectory(roundtrip)
 add_subdirectory(stencil)
 add_subdirectory(threading_test)
 add_subdirectory(toml_test)
+add_subdirectory(toml_reader_test)
 add_subdirectory(utility_formats)
 add_subdirectory(yaml_test)
